Add new proxy_headers argument to ProxyInfo constructor.
This argument lets the user add to or override the headers in the
HTTP CONNECT request sent to the proxy.  Some proxy servers need
additional arguments that are not provided by default by httplib2.
Note that although this commit makes this enhancement for both the
python2 and python3 versions of this library, the python3 version
does not implement ProxyInfo at all (and never has).

Signed-off-by: Martin Carroll <martin.carroll@nokia-bell-labs.com>