Add new proxy_headers argument to ProxyInfo constructor.
This argument lets the user add to or override the headers in the
HTTP CONNECT request sent to the proxy.  Some proxy servers need
additional arguments that are not provided by default by httplib2.
Note that this commit makes this enhancement only for the python2
version of this library, because the python3 version does not
actually implement ProxyInfo.

Signed-off-by: Martin Carroll <martin.carroll@nokia-bell-labs.com>